Store event list view mode in cookie
