+++ apps/web/src/app/arrangementer/hooks/useEventsView.ts
@@ -1,26 +1,35 @@
 "use client"
 
 import { useSearchParams } from "next/navigation"
-import { useMemo } from "react"
+import { useEffect, useMemo, useState } from "react"
+import { setEventsListViewCookie, type EventsListViewMode } from "./eventViewCookie"
 
 export type EventsView = "cards" | "list" | "month" | "week"
 
-export const useEventsView = () => {
+export const useEventsView = (initialListViewMode: EventsListViewMode) => {
   const searchParams = useSearchParams()
+  const [listViewMode, setListViewMode] = useState<EventsListViewMode>(initialListViewMode)
 
   const view = useMemo<EventsView>(() => {
     const viewParameter = searchParams.get("view")
 
-    if (viewParameter === "month" || viewParameter === "week") {
+    if (
+      viewParameter === "month" ||
+      viewParameter === "week" ||
+      viewParameter === "list" ||
+      viewParameter === "cards"
+    ) {
       return viewParameter
     }
 
-    if (viewParameter === "list") {
-      return "list"
-    }
+    return listViewMode
+  }, [searchParams, listViewMode])
 
-    return "cards"
-  }, [searchParams])
+  useEffect(() => {
+    if (view === "list" || view === "cards") {
+      setEventsListViewCookie(view)
+    }
+  }, [view])
 
   return {
     view,
@@ -29,5 +38,6 @@ export const useEventsView = () => {
     isCalendar: view === "month" || view === "week",
     isMonth: view === "month",
     isWeek: view === "week",
+    setListViewMode,
   }
 }
